backend/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
colorama==0.4.6
dlib==19.24.2
etils==1.8.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.110.0
filelock==3.13.3
flatbuffers==24.3.25
flax==0.8.2
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
huggingface-hub==0.22.0
idna==3.6
importlib_resources==6.4.0
jax==0.4.25
jaxlib==0.4.25
Jinja2==3.1.3
keras==3.1.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
namex==0.0.7
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optax==0.2.1
optree==0.11.0
orbax-checkpoint==0.5.7
packaging==24.0
pillow==10.2.0
protobuf==4.25.3
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
safetensors==0.4.2
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
tensorstore==0.1.56
termcolor==2.4.0
tokenizers==0.15.2
toolz==0.12.1
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.39.1
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.18.1
